requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via fsspec
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via aiohttp
certifi==2025.4.26
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via wandb
datasets==3.6.0
    # via -r requirements.in
datasketch==1.6.5
    # via -r requirements.in
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
filelock==3.18.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2025.3.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via wandb
hf-xet==1.1.2
    # via huggingface-hub
huggingface-hub==0.32.1
    # via
    #   datasets
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   yarl
jinja2==3.1.6
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
networkx==3.4.2
    # via torch
numpy==2.2.6
    # via
    #   datasets
    #   datasketch
    #   pandas
    #   scipy
    #   transformers
packaging==25.0
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
    #   wandb
pandas==2.3.0
    # via datasets
platformdirs==4.3.8
    # via wandb
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via wandb
psutil==7.0.0
    # via wandb
pyarrow==20.0.0
    # via datasets
pydantic==2.11.7
    # via wandb
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
    #   wandb
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
    #   wandb
safetensors==0.5.3
    # via transformers
scipy==1.15.3
    # via datasketch
sentry-sdk==2.30.0
    # via wandb
setproctitle==1.3.6
    # via wandb
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sympy==1.14.0
    # via torch
tokenizers==0.21.1
    # via transformers
torch==2.7.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
transformers==4.52.3
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
    #   wandb
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via
    #   requests
    #   sentry-sdk
wandb==0.20.1
    # via -r requirements.in
xxhash==3.5.0
    # via datasets
yarl==1.20.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
